[build] Add MONO_RT_EXTERNAL_ONLY macro to the public API.
This macro should be used to decorate functions that are part of the public API but the runtime is
not supposed to call.

This is to allow us to keep them around but not go crazy with commit reintroducing their usage.


Commit migrated from https://github.com/mono/mono/commit/afa75917e70463663422d22ab4c9a633f930a7af