[sgen] Don't wait for sweep finish when allocating major object
Allocating a major block needs addition to the allocated_blocks array, operation that needs sweep to be finished because sweep moves elements and changes the size when compacting the array. Do this operation in the sweep start instead, when we are not racing.


Commit migrated from https://github.com/mono/mono/commit/abd564253a0dd3d65cdfda24640ca5fc168266d9