[threadpool-io] Fix crash on shutdownin System.Net.Sockets.Socket:Close_internal
There could be a race between System.Net.Sockets.Socket:Close_internal and the io threadpool cleanup, which would lead to a native crash. That's because the io threadpool cleanup would clean the mutex, condition variables as well as close the sockets.

The easiest way to fix that is by cleaning only what's going to stop the runtime from shutting down, and in this case the background selector thread. The same is done for the threadpool where we just shutdown the monitor thread and the worker threads.


Commit migrated from https://github.com/mono/mono/commit/fb6dbac94179c2f908b1f303d3052754a270b713